backend/services/kiwoom_api.py
"""

_�"
"""



from typing import Dict, List, Optional

from datetime import datetime

from decimal import Decimal

import logging

import os



from services.brokerage_connector import BrokerageAPIBase, AccountInfo, StockPrice

from models.trading_schemas import Order, TradeResult



logger = logging.getLogger(__name__)



# Try to import the actual implementation

try:

    from services.kiwoom_api_implementation import KiwoomOpenAPI

    KIWOOM_AVAILABLE = True

except ImportError as e:

    logger.warning(f"Kiwoom OpenAPI implementation not available: {e}")

    KIWOOM_AVAILABLE = False





class KiwoomAPI(BrokerageAPIBase):

    """

    _�"
    """

    

    def __init__(

        self,

        account_number: str,

        user_id: str = "",

        password: str = "",

        cert_password: str = ""

    ):

        """

        Initialize Kiwoom API client

        

        Args:

            account_number: Trading account number

            user_id: Kiwoom user ID (optional for auto-login)

            password: User password (optional)

            cert_password: Certificate password (optional)

        """

        credentials = {

            "account_number": account_number,

            "user_id": user_id,

            "password": password,

            "cert_password": cert_password

        }

        super().__init__(credentials)

        

        # Initialize actual Kiwoom API if available

        self.kiwoom = None

        if KIWOOM_AVAILABLE:

            try:

                self.kiwoom = KiwoomOpenAPI()

                logger.info("Kiwoom OpenAPI initialized successfully")

            except Exception as e:

                logger.error(f"Failed to initialize Kiwoom OpenAPI: {e}")

                self.kiwoom = None

        else:

            logger.warning("Kiwoom OpenAPI not available. Using mock implementation.")

        

    def authenticate(self) -> bool:

        """

        Authenticate with Kiwoom API

        

        Note: Actual implementation requires:

        - CommConnect() call

        - Event loop for OnEventConnect callback

        - Login window interaction

        

        Returns:

            bool: True if authentication successful

            

        Requirements: 11.1

        """

        if self.kiwoom:

            try:

                self.is_authenticated = self.kiwoom.login()

                if self.is_authenticated:

                    logger.info("Kiwoom authentication successful")

                    # Get account list

                    accounts = self.kiwoom.get_account_list()

                    if accounts:

                        logger.info(f"Available accounts: {accounts}")

                        # Use the specified account or first available

                        if self.credentials["account_number"] not in accounts:

                            logger.warning(

                                f"Specified account {self.credentials['account_number']} "

                                f"not found. Using first available: {accounts[0]}"

                            )

                            self.credentials["account_number"] = accounts[0]

                return self.is_authenticated

            except Exception as e:

                logger.error(f"Kiwoom authentication failed: {e}")

                self.is_authenticated = False

                return False

        else:

            logger.warning(

                "Kiwoom API not available. "

                "Requires Windows OS and ActiveX control."

            )

            self.is_authenticated = False

            return False

    

    def get_stock_price(self, symbol: str) -> StockPrice:

        """

        Get real-time stock price information

        

        Note: Actual implementation uses TR code: opt10001 (주식기본_보"
            symbol: Stock code (e.g., "005930")

            

        Returns:

            StockPrice: Current price information

            

        Requirements: 11.2

        """

        self._ensure_authenticated()

        

        if self.kiwoom:

            try:

                data = self.kiwoom.get_stock_price(symbol)

                if data:

                    return StockPrice(

                        symbol=symbol,

                        price=Decimal(str(data.get("current_price", 0))),

                        volume=data.get("volume", 0),

                        open_price=Decimal(str(data.get("open_price", 0))),

                        high_price=Decimal(str(data.get("high_price", 0))),

                        low_price=Decimal(str(data.get("low_price", 0))),

                        timestamp=datetime.now()

                    )

            except Exception as e:

                logger.error(f"Failed to get stock price: {e}")

        

        logger.warning("Kiwoom get_stock_price not available, returning mock data")

        

        # Placeholder return

        return StockPrice(

            symbol=symbol,

            price=Decimal("0"),

            volume=0,

            open_price=Decimal("0"),

            high_price=Decimal("0"),

            low_price=Decimal("0"),

            timestamp=datetime.now()

        )

    

    def get_account_balance(self) -> AccountInfo:

        """

        Get account balance and total assets

        

        Note: Actual implementation uses TR code: opw00001 (_수금상"
        """

        self._ensure_authenticated()

        

        logger.warning("Kiwoom get_account_balance not implemented")

        

        return AccountInfo(

            account_number=self.credentials["account_number"],

            balance=Decimal("0"),

            available_cash=Decimal("0"),

            total_assets=Decimal("0")

        )

        

        # Real implementation would use:

        # self.ocx.dynamicCall("SetInputValue(QString, QString)", "계좌번호", account_number)

        # self.ocx.dynamicCall("CommRqData(...)", "?�수금상?�현??, "opw00001", ...)

    

    def get_account_holdings(self) -> List[Dict]:

        """

        Get list of currently held stocks

        

        Note: Actual implementation uses TR code: opw00018 (계좌_�"
        """

        self._ensure_authenticated()

        

        logger.warning("Kiwoom get_account_holdings not implemented")

        

        return []

        

        # Real implementation would use:

        # self.ocx.dynamicCall("SetInputValue(QString, QString)", "계좌번호", account_number)

        # self.ocx.dynamicCall("CommRqData(...)", "계좌?��??�고", "opw00018", ...)



    def place_order(self, order: Order) -> TradeResult:

        """

        Place a buy or sell order

        

        Note: Actual implementation uses SendOrder() function

        

        Args:

            order: Order details

            

        Returns:

            TradeResult: Order execution result

            

        Requirements: 11.1

        """

        self._ensure_authenticated()

        

        logger.warning("Kiwoom place_order not implemented")

        

        return TradeResult(

            order_id="",

            symbol=order.symbol,

            trade_type=order.trade_type,

            quantity=order.quantity,

            executed_price=Decimal("0"),

            total_amount=Decimal("0"),

            executed_at=datetime.now(),

            status="FAILED",

            message="Kiwoom API not implemented"

        )

        

        # Real implementation would use:

        # order_type = 1 if order.trade_type == "BUY" else 2

        # hoga_type = "00" if order.order_type == "LIMIT" else "03"  # 00: 지_�"
        #     "SendOrder(QString, QString, QString, int, QString, int, int, QString, QString)",

        #     ["주문", "0101", account_number, order_type, order.symbol,

        #      order.quantity, int(order.price) if order.price else 0, hoga_type, ""]

        # )

    

    def cancel_order(self, order_id: str) -> bool:

        """

        Cancel a pending order

        

        Note: Actual implementation uses SendOrder() with order type 3 (취소)

        

        Args:

            order_id: Order identifier

            

        Returns:

            bool: True if cancellation successful

        """

        self._ensure_authenticated()

        

        logger.warning("Kiwoom cancel_order not implemented")

        

        return False

        

        # Real implementation would use:

        # result = self.ocx.dynamicCall(

        #     "SendOrder(...)",

        #     ["취소", "0101", account_number, 3, order.symbol, 0, 0, "00", order_id]

        # )

    

    def get_order_status(self, order_id: str) -> Dict:

        """

        Get status of a specific order

        

        Note: Actual implementation uses TR code: opt10075 (미체결요�?

        

        Args:

            order_id: Order identifier

            

        Returns:

            Dict: Order status information

        """

        self._ensure_authenticated()

        

        logger.warning("Kiwoom get_order_status not implemented")

        

        return {

            "order_id": order_id,

            "status": "UNKNOWN",

            "message": "Kiwoom API not implemented"

        }

        

        # Real implementation would use:

        # self.ocx.dynamicCall("SetInputValue(QString, QString)", "계좌번호", account_number)

        # self.ocx.dynamicCall("CommRqData(...)", "미체결요�?, "opt10075", ...)





# Example usage (for reference):

"""

To use Kiwoom API in production, you would need:



1. Install Kiwoom OpenAPI+ from https://www.kiwoom.com/

2. Install PyQt5: pip install PyQt5

3. Implement event-driven architecture:



from PyQt5.QAxContainer import QAxWidget

from PyQt5.QtCore import QEventLoop



class KiwoomAPI(BrokerageAPIBase):

    def __init__(self, ...):

        super().__init__(...)

        self.ocx = QAxWidget("KHOPENAPI.KHOpenAPICtrl.1")

        self.ocx.OnEventConnect.connect(self._on_event_connect)

        self.ocx.OnReceiveTrData.connect(self._on_receive_tr_data)

        self.ocx.OnReceiveChejanData.connect(self._on_receive_chejan_data)

        

    def _on_event_connect(self, err_code):

        if err_code == 0:

            self.is_authenticated = True

            

    def _on_receive_tr_data(self, screen_no, rqname, trcode, ...):

        # Handle TR response data

        pass

        

    def _on_receive_chejan_data(self, gubun, item_cnt, fid_list):

        # Handle real-time order/balance updates

        pass



4. Run within Qt event loop:



from PyQt5.QtWidgets import QApplication

import sys



app = QApplication(sys.argv)

kiwoom = KiwoomAPI(...)

kiwoom.authenticate()

app.exec_()

"""

