backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# HTTP Client
requests==2.31.0
httpx==0.26.0

# Scheduling
apscheduler==3.10.4

# Environment Variables
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7

# Optional: Redis for caching
redis==5.0.1

# Security
cryptography==42.0.0
pyotp==2.9.0

# Kiwoom Securities API (Windows only)
# PyQt5==5.15.10  # Uncomment for Windows with Kiwoom OpenAPI+

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
